backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
python-dotenv==1.2.2
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.22
google-genai==1.66.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
pdfplumber==0.11.9
pdfminer.six==20251230
pillow==12.1.1
APScheduler==3.11.2
requests==2.32.5
httpx==0.28.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
starlette==0.52.1
anyio==4.12.1
orjson==3.11.7
cryptography==46.0.5
python-slugify==8.0.4
pypdfium2==5.5.0
filetype==1.2.0